Switch up default impl strategy for SIMD methods
Instead of hooking everything up to .process_slice(), make each method
defer to the "next narrowest" variant. This way e.g. if we have a SIMD
impl for .process_chunk128() we get faster impls for
.process_chunk256(), .process_chunk512() for free.

The new strategy requires "splitting arrays in half", which we can do
safely and at zero cost thanks to the awesome arrayref crate.